pyproject.toml
[project]
name = "spectra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "datasets>=4.1.1",
    "evaluate>=0.4.6",
    "hydra-core>=1.3.2",
    "imagecodecs>=2025.3.30",
    "lightning>=2.5.5",
    "numpy>=2.2.6",
    "pycocotools>=2.0.10",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "shapely>=2.1.2",
    "spyder-kernels>=3.0.5",
    "supervision>=0.26.1",
    "tensorboard>=2.20.0",
    "tiffile>=2018.10.18",
    "torchmetrics>=1.8.2",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "ultralytics>=8.3.202",
    "vieweryolo>=0.2.0",
]
